requirements.txt
# Discord библиотека
discord.py>=2.3.0,<3.0.0

# Работа с переменными окружения
python-dotenv>=1.0.0,<2.0.0

# WebSocket для WebRCON
websockets>=12.0,<13.0.0

# MySQL/MariaDB асинхронный драйвер
aiomysql>=0.2.0,<0.3.0

# YAML конфигурация
PyYAML>=6.0,<7.0



